refactor: add products controller and router, search controller and router
