+++ 02-express-tutorial/controllers/people.js
@@ -8,7 +8,7 @@ const addPerson = (req, res) => {
   const { name } = req.body;
 
   if (name) {
-    people.push({ id: people.length, name: req.body.name });
+    people.push({ id: people.length + 1, name: req.body.name });
     res.status(201).json({ success: true, name: req.body.name });
   } else {
     res.status(400).json({ success: false, message: 'Please provide a name' });
