+++ 02-express-tutorial/routes/people.js
@@ -1,14 +1,10 @@
 const express = require('express');
 const router = express.Router();
 
-const { addPerson, getPeople } = require('../controllers/people.js')
+const { addPerson, getPeople } = require('../controllers/people');
 
-router.get('/', (req, res) => {
-  return getPeople(req, res);
-});
-
-router.post('/', (req, res) => {
-  return addPerson(req, res);
-});
+router.route('/').get(getPeople).post(addPerson);
+// router.get('/', getPeople)
+// router.post('/', addPerson)
 
 module.exports = router;
